requirements.txt
streamlit==1.40.1
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.1.2
pytest==7.4.3
pytest-cov==4.1.0
hypothesis==6.92.1
black==23.11.0
pylint==3.0.3
